+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Move APP NAME to constants from .env
 + Upgrade dependencies
 + Refactor dockerfiles
++ Add amount component
 
 ## 1.3.0
 + Move to Next.js framework